Refactor: Move classes to Jetpack package
