+++ WordPress/src/test/java/org/wordpress/android/ui/accounts/login/jetpack/LoginSiteCheckErrorViewModelTest.kt
@@ -1,4 +1,4 @@
-package org.wordpress.android.ui.accounts.login
+package org.wordpress.android.ui.accounts.login.jetpack
 
 import kotlinx.coroutines.InternalCoroutinesApi
 import org.assertj.core.api.Assertions.assertThat
